Win32 port
Former-commit-id: c98281387ed559d0e2db8b44f205e472406806bd [formerly a0f747e00124d8387cce74390e34a61651deb3a9]
Former-commit-id: c5b4bccf3e3d00d28da4d95fab2b47412e3ed7e1

+++ solver/include/procstatus.h
@@ -12,7 +12,7 @@
 #include <iostream>
 #include <limits>
 #include <fstream>
-
+#include <cstdint>
 using namespace std;
 
 /*
@@ -23,17 +23,17 @@ class ProcStatus
 
 private:
 
-	static u_int memlimit;
+	static uint64_t memlimit;
 
 public:
 
-	static u_int maxusedmem;
+	static uint64_t maxusedmem;
 
 	ProcStatus();
 	virtual ~ProcStatus();
 
-	static void setMemLimit( u_int lim );
-	static u_int mem();
+	static void setMemLimit( uint64_t lim );
+	static uint64_t mem();
 	static bool memOK();
 
 };
